make xtensa folder generic for adding more xtensa product varients
